+++ CHANGELOG.md
@@ -14,6 +14,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### New features
 
+* `jj squash` now accepts `--editor` / `-E` to edit the squashed commit message.
+
 ### Fixed bugs
 
 ## [0.35.0] - 2025-11-05